packages/nouns-assets/pyproject.toml
[tool.poetry]
name = "nouns-assets"
version = "0.1.0"
description = ""
authors = ["cardene <chotips08@gmail.com>"]
readme = "README.md"
packages = [{include = "nouns_assets"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
numpy = "^1.26.4"
scikit-learn = "^1.4.1.post1"
matplotlib = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
